# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Loup — household services marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/auth/sso:
    get:
      operationId: startSso
      tags: [auth]
      summary: Start WorkOS SSO for an institution
      description: Redirects the browser to the institution's WorkOS connection.
      parameters:
        - name: slug
          in: query
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Redirect to the WorkOS authorize URL
        "400":
          description: Missing slug
        "404":
          description: Unknown institution
        "503":
          description: SSO not configured (demo mode)
  /v1/auth/callback:
    get:
      operationId: ssoCallback
      tags: [auth]
      summary: WorkOS OAuth callback
      description: Exchanges the authorization code, resolves the profile to a Loup principal, and redirects to the web app with a signed token.
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Redirect to the web app login with token or error
        "400":
          description: Missing code or state
        "503":
          description: SSO not configured (demo mode)
...
  /pack/messages:
    get:
      operationId: listPackMessages
      tags: [pack]
      summary: Household thread messages, oldest first (marks thread read)
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PackMessage"
    post:
      operationId: sendPackMessage
      tags: [pack]
      summary: Post a message to the household thread
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackMessageInput"
      responses:
        "201":
          description: Created message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackMessage"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /pack/requests:
    get:
      operationId: listServiceRequests
      tags: [pack]
      summary: Service requests from household members, newest first
      responses:
        "200":
          description: Requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceRequest"
    post:
      operationId: createServiceRequest
      tags: [pack]
      summary: Request a service for head-of-household approval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRequestInput"
      responses:
        "201":
          description: Created request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /pack/requests/{id}/approve:
    post:
      operationId: approveServiceRequest
      tags: [pack]
      summary: Approve a pending request and create the booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request after approval, with the created booking id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
        "400":
          description: Request is not pending
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /pack/requests/{id}/decline:
    post:
      operationId: declineServiceRequest
      tags: [pack]
      summary: Decline a pending request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request after declining
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
        "400":
          description: Request is not pending
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
components:
  schemas:
...
    ApiMessage:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    PackMessage:
      type: object
      properties:
        id:
          type: integer
        memberId:
          type: integer
        memberName:
          type: string
        initials:
          type: string
        isCurrentUser:
          type: boolean
        body:
          type: string
        sentAt:
          type: string
          format: date-time
      required:
        - id
        - memberId
        - memberName
        - initials
        - isCurrentUser
        - body
        - sentAt
    PackMessageInput:
      type: object
      properties:
        body:
          type: string
          minLength: 1
      required:
        - body
    ServiceRequest:
      type: object
      properties:
        id:
          type: integer
        memberId:
          type: integer
        memberName:
          type: string
        initials:
          type: string
        serviceId:
          type: integer
        serviceName:
          type: string
        providerName:
          type: string
        categoryName:
          type: string
        categoryIcon:
          type: string
        price:
          type: number
        note:
          type: string
        status:
          type: string
          enum: [pending, approved, declined]
        bookingId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        decidedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - memberId
        - memberName
        - initials
        - serviceId
        - serviceName
        - providerName
        - categoryName
        - categoryIcon
        - price
        - note
        - status
        - bookingId
        - createdAt
        - decidedAt
    ServiceRequestInput:
      type: object
      properties:
        serviceId:
          type: integer
        note:
          type: string
          minLength: 1
      required:
        - serviceId
        - note
...